+++ R/modelStudio.R
@@ -1,9 +1,10 @@
 #' @title Interactive Studio for Explanatory Model Analysis
 #'
 #' @description
-#' This function computes various (instance and dataset level) model explanations and produces an interactive,
-#' customisable dashboard. It consists of multiple panels for plots with their short descriptions.
-#' Easily save and share the HTML dashboard with others. Tools for model exploration unite with tools for
+#' This function computes various (instance and dataset level) model explanations and
+#' produces a customisable dashboard, which consists of multiple panels for plots with their
+#' short descriptions. Easily save the dashboard and share it with others. Tools for
+#' \href{https://pbiecek.github.io/ema}{Explanatory Model Analysis} unite with tools for
 #' Exploratory Data Analysis to give a broad overview of the model behavior.
 #'
 #' The extensive documentation covers:
@@ -20,7 +21,8 @@
 #' Displayed variable can be changed by clicking on the bars of plots or with the first dropdown list,
 #'  and observation can be changed with the second dropdown list.
 #' The dashboard gathers useful, but not sensitive, information about how it is being used (e.g. computation length,
-#'  package version, dashboard dimensions). This is for the development purposes only and can be blocked by setting \code{telemetry} to \code{FALSE}.
+#'  package version, dashboard dimensions). This is for the development purposes only and can be blocked
+#'  by setting \code{telemetry} to \code{FALSE}.
 #'
 #' @param explainer An \code{explainer} created with \code{DALEX::explain()}.
 #' @param new_observation New observations with columns that correspond to variables used in the model.
@@ -37,7 +39,7 @@
 #' @param B Number of permutation rounds used for calculation of SV and FI.
 #'  Default is \code{10}.
 #'  See \href{https://modelstudio.drwhy.ai/articles/ms-perks-features.html#more-calculations-means-more-time}{\bold{vignette}}
-#' @param eda Compute EDA plots. Default is \code{TRUE}.
+#' @param eda Compute EDA plots and Residuals vs Feature plot, which adds the data to the dashboard. Default is \code{TRUE}.
 #' @param show_info Verbose a progress on the console. Default is \code{TRUE}.
 #' @param parallel Speed up the computation using \code{parallelMap::parallelMap()}.
 #'  See \href{https://modeloriented.github.io/modelStudio/articles/ms-perks-features.html#parallel-computation}{\bold{vignette}}.
@@ -65,9 +67,9 @@
 #'
 #' \itemize{
 #'   \item The input object is implemented in \href{https://modeloriented.github.io/DALEX/}{\bold{DALEX}}
-#'   \item Feature Importance, Ceteris Paribus, Partial Dependence and Accumulated Dependence plots
+#'   \item Feature Importance, Ceteris Paribus, Partial Dependence and Accumulated Dependence explanations
 #'    are implemented in \href{https://modeloriented.github.io/ingredients/}{\bold{ingredients}}
-#'   \item Break Down and Shapley Values plots are implemented in
+#'   \item Break Down and Shapley Values explanations are implemented in
 #'    \href{https://modeloriented.github.io/iBreakDown/}{\bold{iBreakDown}}
 #' }
 #'
@@ -314,7 +316,10 @@ modelStudio.explainer <- function(explainer,
   ad_data <- prepare_accumulated_dependence(ad_n, ad_c, variables = variable_names)
 
   if (eda) {
-    fd_data <- prepare_feature_distribution(data, y, variables = variable_names)
+    #:# fd_data is used by targetVs and residualsVs plots
+    residuals <- DALEX::model_diagnostics(explainer)$residuals
+    fd_data <- prepare_feature_distribution(data, y, variables = variable_names,
+                                            residuals = residuals)
     at_data <- prepare_average_target(data, y, variables = variable_names)
   } else {
     fd_data <- at_data <- NULL