+++ R/ms_options.R
@@ -33,7 +33,7 @@
 #' }
 #' \subsection{Plot specific options:}{
 #' \code{**} is a two letter code unique to each plot, might be
-#' one of \code{[bd,sv,cp,fi,pd,ad,fd,tv,at]}.\cr
+#' one of \code{[bd,sv,cp,fi,pd,ad,rv,fd,tv,at]}.\cr
 #'
 #' \describe{
 #' \item{**_title}{Plot specific title. Default varies.}
@@ -50,9 +50,10 @@
 #'
 #' \itemize{
 #'   \item The input object is implemented in \href{https://modeloriented.github.io/DALEX/}{\bold{DALEX}}
-#'   \item Feature Importance, Ceteris Paribus, Partial Dependence and Accumulated Dependence plots
-#' are implemented in \href{https://modeloriented.github.io/ingredients/}{\bold{ingredients}}
-#'   \item Break Down and Shapley Values plots are implemented in \href{https://modeloriented.github.io/iBreakDown/}{\bold{iBreakDown}}
+#'   \item Feature Importance, Ceteris Paribus, Partial Dependence and Accumulated Dependence explanations
+#'    are implemented in \href{https://modeloriented.github.io/ingredients/}{\bold{ingredients}}
+#'   \item Break Down and Shapley Values explanations are implemented in
+#'    \href{https://modeloriented.github.io/iBreakDown/}{\bold{iBreakDown}}
 #' }
 #'
 #' @seealso
@@ -159,6 +160,10 @@ ms_options <- function(...) {
     ad_line_size = NULL,
     ad_bar_color = NULL,
     ad_line_color = NULL,
+    rv_title = "Residuals vs Feature",
+    rv_subtitle = NULL,
+    rv_point_size = NULL,
+    rv_point_color = NULL,
     fd_title = "Feature Distribution",
     fd_subtitle = NULL,
     fd_bar_width = NULL,